+++ src/exceptions.cc
@@ -2,7 +2,6 @@
 #include "node_internals.h"
 #include "env.h"
 #include "env-inl.h"
-#include "exceptions.h"
 #include "util.h"
 #include "util-inl.h"
 #include "v8.h"
